+++ src/encoder.rs
@@ -152,22 +152,12 @@ pub struct Encoder<'a, W: Write> {
 }
 
 /// Decoding options, internal type, forwarded to the Writer.
+#[derive(Default)]
 struct Options {
     filter: FilterType,
     adaptive_filter: AdaptiveFilterType,
     sep_def_img: bool,
     validate_sequence: bool,
-}
-
-impl Default for Options {
-    fn default() -> Options {
-        Options {
-            filter: FilterType::default(),
-            adaptive_filter: AdaptiveFilterType::default(),
-            sep_def_img: false,
-            validate_sequence: false,
-        }
-    }
 }
 
 impl<'a, W: Write> Encoder<'a, W> {
@@ -188,7 +178,7 @@ impl<'a, W: Write> Encoder<'a, W> {
     ///
     /// When this method is returns successfully then the images written will be encoded as fdAT
     /// chunks, except for the first image that is still encoded as `IDAT`. You can control if the
-    /// first frame should be treated as an animation frame with [`set_sep_def_img()`].
+    /// first frame should be treated as an animation frame with [`Encoder::set_sep_def_img()`].
     ///
     /// This method returns an error if `num_frames` is 0.
     pub fn set_animated(&mut self, num_frames: u32, num_plays: u32) -> Result<()> {
@@ -221,7 +211,7 @@ impl<'a, W: Write> Encoder<'a, W> {
     /// the animation is often interpreted as a thumbnail.
     ///
     /// This method will return an error when animation control was not configured
-    /// (which is doing by calling [`set_animated`]).
+    /// (which is done by calling [`Encoder::set_animated`]).
     pub fn set_sep_def_img(&mut self, sep_def_img: bool) -> Result<()> {
         if self.info.animation_control.is_some() {
             self.options.sep_def_img = sep_def_img;
@@ -321,7 +311,7 @@ impl<'a, W: Write> Encoder<'a, W> {
     /// Set the fraction of time every frame is going to be displayed, in seconds.
     ///
     /// *Note that this parameter can be set for each individual frame after
-    /// [`write_header`] is called. (see [`Writer::set_frame_delay`])*
+    /// [`Encoder::write_header`] is called. (see [`Writer::set_frame_delay`])*
     ///
     /// If the denominator is 0, it is to be treated as if it were 100
     /// (that is, the numerator then specifies 1/100ths of a second).
@@ -353,9 +343,9 @@ impl<'a, W: Write> Encoder<'a, W> {
     /// replace its region in the output buffer.
     ///
     /// *Note that this parameter can be set for each individual frame after
-    /// [`writer_header`] is called. (see [`Writer::set_blend_op`])*
+    /// [`write_header`] is called. (see [`Writer::set_blend_op`])*
     ///
-    /// See the [`BlendOp`] documentaion for the possible values and their effects.
+    /// See the [`BlendOp`] documentation for the possible values and their effects.
     ///
     /// *Note that for the first frame the two blend modes are functionally
     /// equivalent due to the clearing of the output buffer at the beginning
@@ -386,9 +376,9 @@ impl<'a, W: Write> Encoder<'a, W> {
     /// at the end of the delay (before rendering the next frame)
     ///
     /// *Note that this parameter can be set for each individual frame after
-    /// [`writer_header`] is called (see [`Writer::set_dispose_op`])*
+    /// [`write_header`] is called (see [`Writer::set_dispose_op`])*
     ///
-    /// See the [`DisposeOp`] documentaion for the possible values and their effects.
+    /// See the [`DisposeOp`] documentation for the possible values and their effects.
     ///
     /// *Note that if the first frame uses [`DisposeOp::Previous`]
     /// it will be treated as [`DisposeOp::Background`].*
@@ -462,11 +452,19 @@ impl<'a, W: Write> Encoder<'a, W> {
 /// to guarantee the next image to be prefaced with a fcTL-chunk, and all other chunks would be
 /// guaranteed to be `IDAT`/not affected by APNG's frame control.
 pub struct Writer<W: Write> {
+    /// The underlying writer.
     w: W,
+    /// The local version of the `Info` struct.
     info: PartialInfo,
+    /// Global encoding options.
     options: Options,
-    written: u64,
+    /// The total number of image frames, counting all consecutive IDAT and fdAT chunks.
+    images_written: u64,
+    /// The total number of animation frames, that is equivalent to counting fcTL chunks.
     animation_written: u32,
+    /// A flag to note when the IEND chunk was already added.
+    /// This is only set on code paths that drop `Self` to control the destructor.
+    iend_written: bool,
 }
 
 /// Contains the subset of attributes of [Info] needed for [Writer] to function
@@ -513,15 +511,16 @@ impl PartialInfo {
     /// Converts this partial info to an owned Info struct,
     /// setting missing values to their defaults
     fn to_info(&self) -> Info<'static> {
-        let mut info = Info::default();
-        info.width = self.width;
-        info.height = self.height;
-        info.bit_depth = self.bit_depth;
-        info.color_type = self.color_type;
-        info.frame_control = self.frame_control;
-        info.animation_control = self.animation_control;
-        info.compression = self.compression;
-        info
+        Info {
+            width: self.width,
+            height: self.height,
+            bit_depth: self.bit_depth,
+            color_type: self.color_type,
+            frame_control: self.frame_control,
+            animation_control: self.animation_control,
+            compression: self.compression,
+            ..Default::default()
+        }
     }
 }
 
@@ -544,8 +543,9 @@ impl<W: Write> Writer<W> {
             w,
             info,
             options,
-            written: 0,
+            images_written: 0,
             animation_written: 0,
+            iend_written: false,
         }
     }
 
@@ -602,7 +602,7 @@ impl<W: Write> Writer<W> {
 
         match self.info.animation_control {
             None => {
-                if self.written == 0 {
+                if self.images_written == 0 {
                     Ok(())
                 } else {
                     Err(EncodingError::Format(FormatErrorKind::EndReached.into()))
@@ -623,9 +623,9 @@ impl<W: Write> Writer<W> {
             return Ok(());
         }
 
-        if self.info.animation_control.is_some() && self.info.frame_control.is_some() {
-            Err(EncodingError::Format(FormatErrorKind::MissingFrames.into()))
-        } else if self.written == 0 {
+        if (self.info.animation_control.is_some() && self.info.frame_control.is_some())
+            || self.images_written == 0
+        {
             Err(EncodingError::Format(FormatErrorKind::MissingFrames.into()))
         } else {
             Ok(())
@@ -697,7 +697,7 @@ impl<W: Write> Writer<W> {
                 self.animation_written += 1;
 
                 // If the default image is the first frame of an animation, it's still an IDAT.
-                if self.written == 0 {
+                if self.images_written == 0 {
                     self.write_zlib_encoded_idat(&zlib_encoded)?;
                 } else {
                     let buff_size = zlib_encoded.len().min(Self::MAX_fdAT_CHUNK_LEN as usize);
@@ -712,19 +712,29 @@ impl<W: Write> Writer<W> {
             }
         }
 
-        self.written += 1;
+        self.increment_images_written();
+
+        Ok(())
+    }
+
+    fn increment_images_written(&mut self) {
+        self.images_written = self.images_written.saturating_add(1);
+
         if let Some(actl) = self.info.animation_control {
             if actl.num_frames <= self.animation_written {
                 // If we've written all animation frames, all following will be normal image chunks.
                 self.info.frame_control = None;
             }
         }
+    }
 
-        Ok(())
+    fn write_iend(&mut self) -> Result<()> {
+        self.iend_written = true;
+        self.write_chunk(chunk::IEND, &[])
     }
 
     fn should_skip_frame_control_on_default_image(&self) -> bool {
-        self.options.sep_def_img && self.written == 0
+        self.options.sep_def_img && self.images_written == 0
     }
 
     fn write_zlib_encoded_idat(&mut self, zlib_encoded: &[u8]) -> Result<()> {
@@ -783,7 +793,7 @@ impl<W: Write> Writer<W> {
     /// - The image is not an animated;
     ///
     /// - The selected dimension, considering also the current frame position,
-    ///   goes outside the image boudries;
+    ///   goes outside the image boundaries;
     ///
     /// - One or both the width and height are 0;
     ///
@@ -814,7 +824,7 @@ impl<W: Write> Writer<W> {
     /// - The image is not animated;
     ///
     /// - The selected position, considering also the current frame dimension,
-    ///   goes outside the image boudries;
+    ///   goes outside the image boundaries;
     ///
     // ??? TODO ???
     // - The next frame is the default image
@@ -875,7 +885,7 @@ impl<W: Write> Writer<W> {
     /// into the current output buffer content, or whether it should completely
     /// replace its region in the output buffer.
     ///
-    /// See the [`BlendOp`] documentaion for the possible values and their effects.
+    /// See the [`BlendOp`] documentation for the possible values and their effects.
     ///
     /// *Note that for the first frame the two blend modes are functionally
     /// equivalent due to the clearing of the output buffer at the beginning
@@ -898,7 +908,7 @@ impl<W: Write> Writer<W> {
     /// The dispose operation specifies how the output buffer should be changed
     /// at the end of the delay (before rendering the next frame)
     ///
-    /// See the [`DisposeOp`] documentaion for the possible values and their effects.
+    /// See the [`DisposeOp`] documentation for the possible values and their effects.
     ///
     /// *Note that if the first frame uses [`DisposeOp::Previous`]
     /// it will be treated as [`DisposeOp::Background`].*
@@ -961,17 +971,22 @@ impl<W: Write> Writer<W> {
     /// Unlike a simple drop this ensures that the final chunk was written correctly. When other
     /// validation options (chunk sequencing) had been turned on in the configuration then it will
     /// also do a check on their correctness _before_ writing the final chunk.
-    pub fn finish(self) -> Result<()> {
+    pub fn finish(mut self) -> Result<()> {
         self.validate_sequence_done()?;
-        // Prevent the writing on Drop, we do it fallibly.
-        let mut finalize = mem::ManuallyDrop::new(self);
-        finalize.write_chunk(chunk::IEND, &[])
+        self.write_iend()?;
+        self.w.flush()?;
+
+        // Explicitly drop `self` just for clarity.
+        drop(self);
+        Ok(())
     }
 }
 
 impl<W: Write> Drop for Writer<W> {
     fn drop(&mut self) {
-        let _ = self.write_chunk(chunk::IEND, &[]);
+        if !self.iend_written {
+            let _ = self.write_iend();
+        }
     }
 }
 
@@ -1006,7 +1021,7 @@ impl<'a, W: Write> DerefMut for ChunkOutput<'a, W> {
 /// data into a PNG chunk, based on the image metadata
 ///
 /// Currently the way it works is that the specified buffer
-/// will hold one chunk at the time and bufferize the incoming
+/// will hold one chunk at the time and buffer the incoming
 /// data until `flush` is called or the maximum chunk size
 /// is reached.
 ///
@@ -1033,7 +1048,7 @@ impl<'a, W: Write> ChunkWriter<'a, W> {
         //               is 64 bits.
         const CAP: usize = std::u32::MAX as usize >> 1;
         let curr_chunk;
-        if writer.written == 0 {
+        if writer.images_written == 0 {
             curr_chunk = chunk::IDAT;
         } else {
             curr_chunk = chunk::fdAT;
@@ -1077,7 +1092,7 @@ impl<'a, W: Write> ChunkWriter<'a, W> {
         assert_eq!(self.index, 0, "Called when not flushed");
         let wrt = self.writer.deref_mut();
 
-        self.curr_chunk = if wrt.written == 0 {
+        self.curr_chunk = if wrt.images_written == 0 {
             chunk::IDAT
         } else {
             chunk::fdAT
@@ -1101,7 +1116,7 @@ impl<'a, W: Write> ChunkWriter<'a, W> {
     /// as it is updated internally.
     fn set_fctl(&mut self, f: FrameControl) {
         if let Some(ref mut fctl) = self.writer.info.frame_control {
-            // ingnore the sequence number
+            // Ignore the sequence number
             *fctl = FrameControl {
                 sequence_number: fctl.sequence_number,
                 ..f
@@ -1147,7 +1162,7 @@ impl<'a, W: Write> Write for ChunkWriter<'a, W> {
             }
         }
 
-        // cap the buffer length to the maximum nuber of bytes that can't still
+        // Cap the buffer length to the maximum number of bytes that can't still
         // be added to the current chunk
         let written = data.len().min(self.buffer.len() - self.index);
         data = &data[..written];
@@ -1185,7 +1200,7 @@ impl<W: Write> Drop for ChunkWriter<'_, W> {
 /// exiting the function, and this is where `Wrapper` comes in.
 ///
 /// Unfortunately the `ZlibWriter` can't be used because on the
-/// write following the error, `finish` wuold be called and that
+/// write following the error, `finish` would be called and that
 /// would write some data even if 0 bytes where compressed.
 ///
 /// If the `finish` function fails then there is nothing much to
@@ -1326,7 +1341,7 @@ impl<'a, W: Write> StreamWriter<'a, W> {
     /// - The image is not an animated;
     ///
     /// - The selected dimension, considering also the current frame position,
-    ///   goes outside the image boudries;
+    ///   goes outside the image boundaries;
     ///
     /// - One or both the width and height are 0;
     ///
@@ -1355,7 +1370,7 @@ impl<'a, W: Write> StreamWriter<'a, W> {
     /// - The image is not animated;
     ///
     /// - The selected position, considering also the current frame dimension,
-    ///   goes outside the image boudries;
+    ///   goes outside the image boundaries;
     ///
     pub fn set_frame_position(&mut self, x: u32, y: u32) -> Result<()> {
         if let Some(ref mut fctl) = self.fctl {
@@ -1414,7 +1429,7 @@ impl<'a, W: Write> StreamWriter<'a, W> {
     /// into the current output buffer content, or whether it should completely
     /// replace its region in the output buffer.
     ///
-    /// See the [`BlendOp`] documentaion for the possible values and their effects.
+    /// See the [`BlendOp`] documentation for the possible values and their effects.
     ///
     /// *Note that for the first frame the two blend modes are functionally
     /// equivalent due to the clearing of the output buffer at the beginning
@@ -1437,7 +1452,7 @@ impl<'a, W: Write> StreamWriter<'a, W> {
     /// The dispose operation specifies how the output buffer should be changed
     /// at the end of the delay (before rendering the next frame)
     ///
-    /// See the [`DisposeOp`] documentaion for the possible values and their effects.
+    /// See the [`DisposeOp`] documentation for the possible values and their effects.
     ///
     /// *Note that if the first frame uses [`DisposeOp::Previous`]
     /// it will be treated as [`DisposeOp::Background`].*
@@ -1465,9 +1480,8 @@ impl<'a, W: Write> StreamWriter<'a, W> {
         // TODO: call `writer.finish` somehow?
         self.flush()?;
 
-        match self.writer.take() {
-            Wrapper::Chunk(wrt) => wrt.writer.validate_sequence_done()?,
-            _ => {}
+        if let Wrapper::Chunk(wrt) = self.writer.take() {
+            wrt.writer.validate_sequence_done()?;
         }
 
         Ok(())
@@ -1482,7 +1496,7 @@ impl<'a, W: Write> StreamWriter<'a, W> {
             Wrapper::Chunk(wrt) => wrt,
             Wrapper::Unrecoverable => {
                 let err = FormatErrorKind::Unrecoverable.into();
-                return Err(EncodingError::Format(err).into());
+                return Err(EncodingError::Format(err));
             }
             Wrapper::Zlib(_) => unreachable!("never called on a half-finished frame"),
             Wrapper::None => unreachable!(),
@@ -1496,8 +1510,9 @@ impl<'a, W: Write> StreamWriter<'a, W> {
         let (scansize, size) = wrt.next_frame_info();
         self.line_len = scansize;
         self.to_write = size;
-        wrt.writer.written += 1;
+
         wrt.write_header()?;
+        wrt.writer.increment_images_written();
 
         // now it can be taken because the next statements cannot cause any errors
         match self.writer.take() {
@@ -2214,6 +2229,40 @@ mod tests {
         Ok(())
     }
 
+    #[test]
+    fn finish_drops_inner_writer() -> Result<()> {
+        struct NoWriter<'flag>(&'flag mut bool);
+
+        impl Write for NoWriter<'_> {
+            fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+                Ok(buf.len())
+            }
+            fn flush(&mut self) -> io::Result<()> {
+                Ok(())
+            }
+        }
+        impl Drop for NoWriter<'_> {
+            fn drop(&mut self) {
+                *self.0 = true;
+            }
+        }
+
+        let mut flag = false;
+
+        {
+            let mut encoder = Encoder::new(NoWriter(&mut flag), 10, 10);
+            encoder.set_depth(BitDepth::Eight);
+            encoder.set_color(ColorType::Grayscale);
+
+            let mut writer = encoder.write_header()?;
+            writer.write_image_data(&vec![0; 100])?;
+            writer.finish()?;
+        }
+
+        assert!(flag, "PNG finished but writer was not dropped");
+        Ok(())
+    }
+
     /// A Writer that only writes a few bytes at a time
     struct RandomChunkWriter<R: Rng, W: Write> {
         rng: R,
@@ -2223,7 +2272,7 @@ mod tests {
     impl<R: Rng, W: Write> Write for RandomChunkWriter<R, W> {
         fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
             // choose a random length to write
-            let len = cmp::min(self.rng.gen_range(1, 50), buf.len());
+            let len = cmp::min(self.rng.gen_range(1..50), buf.len());
 
             self.w.write(&buf[0..len])
         }