+++ src/lib.rs
@@ -13,8 +13,7 @@
 //! ```
 //! use std::fs::File;
 //! // The decoder is a build for reader and can be used to set various decoding options
-//! // via `Transformations`. The default output transformation is `Transformations::EXPAND
-//! // | Transformations::STRIP_ALPHA`.
+//! // via `Transformations`. The default output transformation is `Transformations::IDENTITY`.
 //! let decoder = png::Decoder::new(File::open("tests/pngsuite/basi0g01.png").unwrap());
 //! let mut reader = decoder.read_info().unwrap();
 //! // Allocate the output buffer.
@@ -60,7 +59,7 @@
 //! ```
 //!
 
-#![deny(unsafe_code)]
+#![forbid(unsafe_code)]
 
 #[macro_use]
 extern crate bitflags;